Refactored the user authentication middleware Moved the request interface to the utils folder Moved the jwt verify functionality to the jwt foler in the utils folder Added the custom error to its own folder close #6
