+++ server/utils/index.ts
@@ -1,0 +1,9 @@
+import CustomError from './CustomError';
+import CustomRequest from './interfaces';
+import verifyToken from './jwt';
+
+export {
+  CustomError,
+  CustomRequest,
+  verifyToken,
+};
