+++ server/utils/interfaces/index.ts
@@ -1,0 +1,3 @@
+import CustomRequest from './requestInterface';
+
+export default CustomRequest;
