+++ server/utils/interfaces/requestInterface.ts
@@ -1,0 +1,7 @@
+import { Request } from 'express';
+
+interface CustomRequest extends Request {
+  user: object | unknown;
+}
+
+export default CustomRequest;
